LPS-95166 Move integration test classes of the com.liferay.portal.service package to module
